requirements.txt
amqp==5.3.1
arrow==1.3.0
asgiref==3.8.1
billiard==4.2.1
binaryornot==0.4.4
boto3==1.36.22
botocore==1.36.22
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cookiecutter==2.6.0
crispy-bootstrap5==2024.10
crispy-tailwind==1.0.3
cron-descriptor==1.4.5
cryptography==44.0.0
Django==5.1.4
django-allauth==65.3.0
django-axes==7.0.1
django-browser-reload==1.17.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-countries==7.6.1
django-crispy-forms==2.3
django-cte-forest==0.2.2
django-debug-toolbar==4.4.6
django-encrypted-model-fields==0.6.5
django-extensions==3.2.3
django-jazzmin==3.0.1
django-js-asset==3.0.1
django-mptt==0.16.0
django-phonenumber-field==8.0.0
django-redis==5.4.0
django-storages==1.14.5
django-suit==0.2.28
django-tailwind==3.8.0
django-timezone-field==7.0
django-treebeard==4.7.1
factory_boy==3.3.1
Faker==33.1.0
gunicorn==23.0.0
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.5
jmespath==1.0.1
kombu==5.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==24.2
paypalrestsdk==1.13.3
phonenumbers==8.13.52
pillow==11.0.0
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.18.0
pyOpenSSL==24.3.0
pytest==8.3.4
pytest-django==4.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
rich==13.9.4
s3transfer==0.11.2
sentry-sdk==2.19.2
six==1.17.0
sqlparse==0.5.3
stripe==11.3.0
text-unidecode==1.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.8.2
